Move RendererHandler into a separate file
